Revert "Merge branch 'Item_feature' into develop"
This reverts commit ba794d4bb43fafc0c2ed3f9c7fc74caee2d04609, reversing
changes made to fc0f6a984837e8789d6bd910dc5b06acc0ecdc47.

+++ Source/Promiss9/Public/P9WeaponData.h
@@ -32,7 +32,7 @@ struct FP9WeaponData : public FTableRowBase
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Weapon")
 	float Damage;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Weapon")
-	float Range;
+	int32 Range;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Weapon")
 	int32 Count;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Weapon")
